refactor using consistent conventional naming conventions for starters. renamed suffix 'api' to 'platform'
