+++ embabel-agent-autoconfigure/src/main/java/com/embabel/agent/autoconfigure/platform/DefaultAgentAutoConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.agent.autoconfigure.defaultconfig;
+package com.embabel.agent.autoconfigure.platform;
 
 
 import com.embabel.agent.config.AgentPlatformConfiguration;
