+++ embabel-agent-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.embabel.agent.autoconfigure.defaultconfig.DefaultAgentAutoConfiguration
+com.embabel.agent.autoconfigure.platform.DefaultAgentAutoConfiguration
