+++ embabel-agent-starters/embabel-agent-starter/pom.xml
@@ -17,7 +17,6 @@
         <dependency>
             <groupId>com.embabel.agent</groupId>
             <artifactId>embabel-agent-autoconfigure</artifactId>
-            <version>${project.version}</version>
         </dependency>
     </dependencies>
 </project>
